Actualizando documentación

+++ install.sh
@@ -1,5 +1,10 @@
 cd ~
-mv .vimrc .vimrc.bak
+
+if [ -f .vimrc ];
+then
+    mv .vimrc .vimrc.bak
+fi
+
 echo 'set runtimepath+=~/.vimcfg
 source ~/.vimcfg/config.vim
 ' > ~/.vimrc
